+++ src/features/WordFeatures.h
@@ -20,7 +20,7 @@
 class CWordFeatures: public CSimpleFeatures<WORD>
 {
 	public:
-		CWordFeatures(LONG size, INT num_symbols=1<<16);
+		CWordFeatures(INT size, INT num_symbols=1<<16);
 		CWordFeatures(const CWordFeatures & orig);
 
 		/** load features from file
